Migrate async cursor tests
