+++ tests/Agent/IntegrationTests/SharedApplications/Common/MultiFunctionApplicationHelpers/MultiFunctionApplicationHelpers.csproj
@@ -44,12 +44,12 @@
     <!-- MongoDB.Driver .NET Framework references -->
     <PackageReference Include="MongoDB.Driver" Version="2.13.1" Condition="'$(TargetFramework)' == 'net462'" />
     <PackageReference Include="MongoDB.Driver" Version="2.13.1" Condition="'$(TargetFramework)' == 'net471'" />
-    <PackageReference Include="MongoDB.Driver" Version="2.13.1" Condition="'$(TargetFramework)' == 'net48'" />
+    <PackageReference Include="MongoDB.Driver" Version="2.17.1" Condition="'$(TargetFramework)' == 'net48'" />
 
     <!-- MongoDB.Driver .NET/Core references -->
     <PackageReference Include="MongoDB.Driver" Version="2.13.1" Condition="'$(TargetFramework)' == 'netcoreapp3.1'" />
     <PackageReference Include="MongoDB.Driver" Version="2.13.1" Condition="'$(TargetFramework)' == 'net5.0'" />
-    <PackageReference Include="MongoDB.Driver" Version="2.13.1" Condition="'$(TargetFramework)' == 'net6.0'" />
+    <PackageReference Include="MongoDB.Driver" Version="2.17.1" Condition="'$(TargetFramework)' == 'net6.0'" />
 
     <!-- MySqlConnector framework references -->
     <PackageReference Include="MySqlConnector" Version="1.0.1" Condition="'$(TargetFramework)' == 'net462'" />
